Add support for `desktop.goto(url)` action (#18)
This patch adds support for the `desktop.goto(url)` action, which opens
the url in a new Firefox tab using the
[`-new-tab`](https://wiki.mozilla.org/Firefox/CommandLineOptions#-new-tab_URL)
flag. Note that we use the `firefox-esr` command since the Dockerfile
installs the firefox-esr package.

Manually visiting a URL can be useful when:
- we already know the starting condition (e.g. "start on amazon.com")
and want to shave a few seconds
- we want to save tokens spent on visiting a known URL